refactor template resolver to components dir
